Remove the AbstractDatabaseCommand.
The functionality is provided via an utility class.
