+++ zav.discord.blanc.runtime/src/main/java/zav/discord/blanc/runtime/command/mod/ResponseRemoveCommand.java
@@ -1,21 +1,26 @@
 package zav.discord.blanc.runtime.command.mod;
 
 import jakarta.persistence.EntityManager;
+import jakarta.persistence.EntityManagerFactory;
 import java.util.EnumSet;
 import java.util.Set;
 import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.interaction.SlashCommandEvent;
 import zav.discord.blanc.api.cache.AutoResponseCache;
+import zav.discord.blanc.command.AbstractGuildCommand;
 import zav.discord.blanc.command.GuildCommandManager;
 import zav.discord.blanc.databind.AutoResponseEntity;
 import zav.discord.blanc.databind.GuildEntity;
+import zav.discord.blanc.runtime.internal.PersistenceUtils;
 
 /**
  * This command allows the user to remove automatic responses. The responses are identified by their
  * (fixed) index in the database.
  */
-public class ResponseRemoveCommand extends AbstractDatabaseCommand {
+public class ResponseRemoveCommand extends AbstractGuildCommand {
   private final AutoResponseCache cache;
+  private final SlashCommandEvent event;
+  private final EntityManagerFactory factory;
   
   /**
    * Creates a new instance of this command.
@@ -24,12 +29,18 @@ public class ResponseRemoveCommand extends AbstractDatabaseCommand {
    * @param manager The manager instance for this command.
    */
   public ResponseRemoveCommand(SlashCommandEvent event, GuildCommandManager manager) {
-    super(event, manager);
-    this.cache = client.getAutoResponseCache();
+    super(manager);
+    this.event = event;
+    this.factory = manager.getClient().getEntityManagerFactory();
+    this.cache = manager.getClient().getAutoResponseCache();
   }
 
   @Override
-  protected String modify(EntityManager entityManager, GuildEntity entity) {
+  public void run() {
+    PersistenceUtils.handle(factory, event, this::modify);
+  }
+
+  private String modify(EntityManager entityManager, GuildEntity entity) {
     int index = (int) event.getOption("index").getAsLong();
     
     if (index < 0 || index >= entity.getAutoResponses().size()) {
@@ -40,7 +51,7 @@ public class ResponseRemoveCommand extends AbstractDatabaseCommand {
     entity.remove(responseEntity);
     
     // Remove the corresponding entry from cache
-    cache.invalidate(guild);
+    cache.invalidate(event.getGuild());
     
     return getMessage("response_removed", responseEntity.getPattern());
   }
